+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Breaking changes in the generated code will be considered as breaking changes wh
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### unreleased
+### 0.5.0 - 2023-10-28
 ### Added
 * Added `create_shader_module_embedded` for including the source as a string literal instead of using `include_str!`.
 * Added `#[derive(Debug)]` to generated types for easier debugging and profiling.